requirements.txt

azure-functions
boto3
pyodbc==4.0.30
msrest==0.6.18
msrestazure==0.6.4
azure-core==1.8.0
azure-common==1.1.25
azure-cosmos==4.0.0b6
azure-identity==1.4.0
azure-mgmt-sql==0.20.0
azure-mgmt-core==1.2.0
azure-mgmt-compute==13.0.0
azure-storage-blob==12.4.0
azure-keyvault-secrets==4.2.0